# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reasoning/assessments/{assessmentId}:
    get:
      operationId: getReasoningAssessment
      tags: [reasoning]
      summary: Get one assessment with its items (no scoring keys exposed)
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAssessment"
  /reasoning/assessments/{assessmentId}/start:
    post:
      operationId: startReasoningAttempt
      tags: [reasoning]
      summary: Start (or resume) an attempt on a diagnostic assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartReasoningBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAttemptState"
  /reasoning/assessments/{assessmentId}/submit:
    post:
      operationId: submitReasoningAttempt
      tags: [reasoning]
      summary: Submit responses; records a pass and returns written feedback
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReasoningBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningResult"
...
components:
  schemas:
...
    ReasoningItem:
      type: object
      required: [id, position, prompt]
      properties:
        id: { type: integer }
        position: { type: integer }
        prompt: { type: string }
        options:
          type: ["array", "null"]
          items: { type: string }
          description: >-
            The answer choices, for the mcq and hybrid formats. Null for the
            written format, where no options are shown.
    ReasoningAssessment:
      type: object
      required: [id, instrument, phase, format, title, instructions, items]
      properties:
        id: { type: integer }
        instrument:
          type: string
          enum: [ethical, critical]
        phase:
          type: string
          enum: [baseline, unit1, unit2, unit3, unit4]
        format:
          type: string
          enum: [mcq, hybrid, written]
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        instructions: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningItem"
    ReasoningAttemptState:
      type: object
      required: [id, assessmentId, status, startedAt, items]
      properties:
        id: { type: integer }
        assessmentId: { type: integer }
        status:
          type: string
          enum: [in_progress, submitted]
        startedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
        passed: { type: ["boolean", "null"] }
        feedback: { type: ["string", "null"] }
        headline:
          type: ["string", "null"]
          description: For a submitted attempt being reviewed — the score headline.
        metrics:
          type: ["array", "null"]
          description: For a submitted attempt being reviewed — the score metrics.
          items:
            $ref: "#/components/schemas/ReasoningMetric"
        review:
          type: ["array", "null"]
          description: For a submitted attempt being reviewed — per-question review with the student's answer and the correct answer.
          items:
            $ref: "#/components/schemas/ReasoningReviewItem"
        items:
          type: array
          description: >-
            The exact items to present for THIS attempt. The first take uses the
            seeded template; each retake returns freshly generated questions of
            the same kind (same instrument, skill areas, and structure).
          items:
            $ref: "#/components/schemas/ReasoningItem"
    ReasoningResponseInput:
      type: object
      required: [itemId]
      properties:
        itemId: { type: integer }
        selectedIndex:
          type: ["integer", "null"]
          description: mcq / hybrid — chosen option index.
        writtenAnswer:
          type: ["string", "null"]
          description: written / hybrid — the student's short written answer.
...
    StartReasoningBody:
      type: object
      properties:
        retake:
          type: boolean
          description: >-
            When true, begin a fresh attempt even if a previous attempt was
            already submitted. An in-progress attempt is still resumed.
    SubmitReasoningBody:
      type: object
      required: [responses]
      properties:
        responses:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningResponseInput"
    ReasoningMetric:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
        detail: { type: ["string", "null"] }
    ReasoningReviewItem:
      type: object
      required: [itemId, prompt]
      properties:
        itemId: { type: integer }
        prompt: { type: string }
        options:
          type: ["array", "null"]
          items: { type: string }
          description: mcq / hybrid — the answer choices shown.
        selectedIndex:
          type: ["integer", "null"]
          description: mcq / hybrid — the option index the student chose.
        correctIndex:
          type: ["integer", "null"]
          description: mcq / hybrid — the correct option index.
        isCorrect:
          type: ["boolean", "null"]
          description: Whether the student's answer was judged correct.
        writtenAnswer:
          type: ["string", "null"]
          description: written / hybrid — the student's short written answer.
        modelAnswer:
          type: ["string", "null"]
          description: written / hybrid — the reference model answer the response was graded against.
        rationale:
          type: ["string", "null"]
          description: written — a short note on why the answer was or wasn't judged correct.
    ReasoningResult:
      type: object
      required: [attemptId, passed, feedback, headline, metrics, review]
      properties:
        attemptId: { type: integer }
        passed: { type: boolean }
        feedback: { type: string }
        headline: { type: string }
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningMetric"
        review:
          type: array
          description: Per-question review — each item with the student's answer and the correct answer.
          items:
            $ref: "#/components/schemas/ReasoningReviewItem"
...