# lib/api-zod/src/generated/types/reasoningAssessment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasoningAssessmentFormat } from './reasoningAssessmentFormat';
import type { ReasoningAssessmentInstrument } from './reasoningAssessmentInstrument';
import type { ReasoningAssessmentPhase } from './reasoningAssessmentPhase';
import type { ReasoningItem } from './reasoningItem';

export interface ReasoningAssessment {
  id: number;
  instrument: ReasoningAssessmentInstrument;
  phase: ReasoningAssessmentPhase;
  format: ReasoningAssessmentFormat;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  instructions: string;
  items: ReasoningItem[];
}