# lib/api-zod/src/generated/types/reasoningItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningItem {
  id: number;
  position: number;
  prompt: string;
  /**
     * The answer choices, for the mcq and hybrid formats. Null for the written format, where no options are shown.
     * @nullable
     */
  options?: string[] | null;
}