# lib/api-zod/src/generated/types/reasoningResponseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningResponseInput {
  itemId: number;
  /**
     * mcq / hybrid — chosen option index.
     * @nullable
     */
  selectedIndex?: number | null;
  /**
     * written / hybrid — the student's short written answer.
     * @nullable
     */
  writtenAnswer?: string | null;
}