# lib/api-zod/src/generated/types/reasoningReviewItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningReviewItem {
  itemId: number;
  prompt: string;
  /**
     * mcq / hybrid — the answer choices shown.
     * @nullable
     */
  options?: string[] | null;
  /**
     * mcq / hybrid — the option index the student chose.
     * @nullable
     */
  selectedIndex?: number | null;
  /**
     * mcq / hybrid — the correct option index.
     * @nullable
     */
  correctIndex?: number | null;
  /**
     * Whether the student's answer was judged correct.
     * @nullable
     */
  isCorrect?: boolean | null;
  /**
     * written / hybrid — the student's short written answer.
     * @nullable
     */
  writtenAnswer?: string | null;
  /**
     * written / hybrid — the reference model answer the response was graded against.
     * @nullable
     */
  modelAnswer?: string | null;
  /**
     * written — a short note on why the answer was or wasn't judged correct.
     * @nullable
     */
  rationale?: string | null;
}